requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiohttp_request==0.0.2
aiosignal==1.4.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
coverage==7.11.3
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
loguru==0.7.3
MarkupSafe==3.0.3
multidict==6.7.0
mypy==1.18.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
pip-review==1.3.0
pluggy==1.6.0
propcache==0.4.1
Pygments==2.19.2
pytest==9.0.0
pytest-asyncio==1.3.0
pytest-cov==7.0.0
python-dotenv==1.2.1
python-telegram-bot==22.5
requests==2.32.5
ruff==0.7.1
setuptools==80.9.0
sniffio==1.3.1
types-requests==2.32.4.20250913
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.22.0
